+++ src/TRegx/CleanRegex/Match/Details/Match.php
@@ -10,6 +10,8 @@ interface Match extends Details
 {
     public function text(): string;
 
+    public function textLength(): int;
+
     public function parseInt(): int;
 
     public function isInt(): bool;