+++ test/Integration/TRegx/CleanRegex/Match/Details/MatchImplTest.php
@@ -8,7 +8,6 @@ use TRegx\CleanRegex\Internal\Match\MatchAll\EagerMatchAllFactory;
 use TRegx\CleanRegex\Internal\Match\UserData;
 use TRegx\CleanRegex\Internal\Model\Adapter\RawMatchesToMatchAdapter;
 use TRegx\CleanRegex\Internal\Model\Matches\RawMatchesOffset;
-use TRegx\CleanRegex\Internal\SubjectableEx;
 use TRegx\CleanRegex\Internal\SubjectableImpl;
 use TRegx\CleanRegex\Match\Details\Match;
 use TRegx\CleanRegex\Match\Details\MatchImpl;
@@ -20,12 +19,14 @@ class MatchImplTest extends TestCase
     const INDEX_MARLA_SINGER = 1;
     const INDEX_ROBERT_PAULSON = 2;
     const INDEX_JACK_SPARROW = 3;
+    const INDEX_EDWARD = 4;
 
     const subject = 'people are always asking me if I know Tyler Durden. and suddenly I realize that all of this: ' . PHP_EOL
     . 'the gun, the bombs, the revolution... has got something to do with a girl named Marla Singer. ' . PHP_EOL
     . 'in death a member of project mayhem has a name. his name is Robert P***son.' . PHP_EOL
     . PHP_EOL
-    . "when you marooned me on that god forsaken spit of land, you forgot one very important thing mate. i'm captain Jack Sparrow.";
+    . "when you marooned me on that god forsaken spit of land, you forgot one very important thing mate. i'm captain Jack Sparrow." . PHP_EOL
+    . "Ędward.";
 
     /**
      * @test
@@ -127,6 +128,21 @@ class MatchImplTest extends TestCase
 
         // then
         $this->assertEquals('Jack Sparrow', $text);
+    }
+
+    /**
+     * @test
+     */
+    public function shouldGetMatchLength()
+    {
+        // given
+        $match = $this->getMatch(self::INDEX_EDWARD);
+
+        // when
+        $length = $match->textLength();
+
+        // then
+        $this->assertEquals(6, $length);
     }
 
     /**
@@ -288,7 +304,7 @@ class MatchImplTest extends TestCase
         $all = $match->all();
 
         // then
-        $this->assertEquals(['Tyler Durden', 'Marla Singer', 'Robert', 'Jack Sparrow'], $all);
+        $this->assertEquals(['Tyler Durden', 'Marla Singer', 'Robert', 'Jack Sparrow', 'Ędward'], $all);
     }
 
     /**
@@ -303,7 +319,7 @@ class MatchImplTest extends TestCase
         $all = $match->group('surname')->all();
 
         // then
-        $this->assertEquals(['Durden', 'Singer', null, 'Sparrow'], $all);
+        $this->assertEquals(['Durden', 'Singer', null, 'Sparrow', null], $all);
     }
 
     /**
@@ -360,7 +376,7 @@ class MatchImplTest extends TestCase
 
     private function getMatch(int $index): Match
     {
-        $pattern = '/(?<firstName>(?<initial>[A-Z])[a-z]+)(?: (?<surname>[A-Z][a-z]+))?/';
+        $pattern = '/(?<firstName>(?<initial>\p{Lu})[a-z]+)(?: (?<surname>[A-Z][a-z]+))?/u';
         preg::match_all($pattern, self::subject, $matches, PREG_OFFSET_CAPTURE);
 
         $rawMatches = new RawMatchesOffset($matches);